Barrels and levers can be tethered and pulled by the thread
Barrels and levers can be tethered and 'flung'.
Flinging a tethered barrel will teleport it to the mouse_pos. Flinging a lever will flip it if pulled in the correct direction (i.e., when the lever leans right it must be pulled left to flip and visa-versa

the fling action is bound to "t". The entity that has the tethered status effect (both in its status_holder & as a global group of the entity)

BUG/TODO: status node is not being removed when flung