Refactored the status_tether
I have refactored the status tethered functions so that bodies that can be tethered are CharacterBody2D's which extend a class "TetherableBody". This allows us to somewhat easily add new bodies that can interact with the player's yarn mechanics.

The following are steps to make a body that can be tethered and how to override the default tethered functions (i.e., pull() & fling())

1. Create a CharacterBody2D (we'll call this foo_CharBody) with the group "tetherable_body".
2. In the foo_CharBody.gd script, replace "extends CharacterBody2D" with "extends TetherableBody"
3. Set the foo_CharBody's collision layers to "tetherable" or 9.
4. Set the foo_CharBody's Z index to 1, and Y sorting to True.
5. Add a child Area2D node to foo_CharBody, name it "TetherableArea2D" for consistency.
6. Set the TetherableArea2D's Collision Layer to "Tetherable" (and the Collision Mask to nothing).
7. Set the TetherableArea2D Input Pickable to true (likely is by default).
7. Add a child Shape2D node to TetherableArea2D with new shape resource "CircleShape2D" with a radius of 12px.
9. Attach TetherableArea2D singals: mouse_entered() and mouse_exited() to the TetherableArea2D.gd script.
10. In the "func _..._mouse_entered()" block type "select()".
11. In the "func _..._mouse_exited()" block type "deselect()".

Lastly to override the TetheredBody functions pull() (Between self and another body) and fling() (Between self and player), write an identical function in the foo_CharBody.gd script.