yarn object with rope physics
Created a player XSM for the crochet mechanic.
Created a yarn segment scene.
Created a yarn object scene and added it to the player.
TODO:
Remove gravity from the yarn object to allow the yarn to lay at rest.
Rework yarn casting/recalling so it rope grows from the player and retracts to the player