Fixes #206
There was some weirdness with the order of which we were saving/loading the inventory, and how. Because, the inventory resource was still loaded in memory despite its parent node being gone.

So, I added a reference to the resource in GameManager that's agnostic of the node. That should fix it from my brief testing.